4 LiFE
Alright. I emailed my biology teacher regarding the score that I didn't receive(yet). I'll still be waiting for the score, so yeah. This commit message will only signify that it has been written through the commit on GitHub servers.

Anyways, I have created the 404 route, allowing for errors to be passed & as well as masking the admin route to redirect to the 404 route, as well as fixing the login required thing from yesterday. Now the admin routes will be updated soon to reflect that it is possible to connect & create the node.

Regards,
DF.
aviance.

Commit made on November 20th, 2022 at 11:15AM PST.